backend/requirements.txt
altgraph==0.17.5
asgiref==3.11.1
charset-normalizer==3.4.4
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
mysqlclient==2.2.8
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
psycopg2-binary==2.9.11
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pywin32-ctypes==0.2.3
reportlab==4.4.10
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
# Agregar estas dependencias al requirements.txt existente
gunicorn==21.2.0
whitenoise==6.7.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9